+++ src/test/java/org/sbrubbles/conditio/ScopeTest.java
@@ -54,7 +54,7 @@ public class ScopeTest {
   public void getHandlersIsUnmodifiable() {
     try (Scope a = Scope.create()) {
       final List<Handler> hs = a.getHandlers();
-      final Handler h = new HandlerImpl(BasicCondition.class, (c, s) -> "test");
+      final Handler h = new HandlerImpl(BasicCondition.class, (c, ops) -> ops.use("test"));
 
       assertThrows(UnsupportedOperationException.class, () -> hs.add(h));
       assertThrows(UnsupportedOperationException.class, () -> hs.remove(h));